fixed bugs in task 4

+++ 0x0B-malloc_free/4-free_grid.c
@@ -9,9 +9,8 @@
 
 void free_grid(int **grid, int height)
 {
-unsigned int i;
-for (i = 0; i < (unsigned int)height; i++)
+int i;
+for (i = 0; i < height; i++)
 	free(grid[i]);
-
 free(grid);
 }
